Update patch set 1
Patch Set 1:

Codeworkx it seems it won't boot past the
CM boot animation.  It just hangs never actually
make it in.

I meant to turn off secure adb before I built and spaced
it or I would have a logcat for you.  Building without
it real fast to make sure that was the cause.

Patch-set: 1