Update patch set 1
Patch Set 1: Code-Review-1

ok -1 for now, Codeworkx, I was able to
pull the part of the logcat log related
to it crashing (the driver) it just stays
on the boot animation.

Sorry about the slightly weird format, I had
to pull it from the raw console as adb didn't
want to respond to commands except for raw output.

http://d-h.st/I8n

Patch-set: 1