Update patch set 1
Patch Set 1: -Code-Review

It still looks good on the 7 inch, Ji said it was disabled on the 10.1 because im sure it starts to look weird at that point.

Eitherway, the option is there in settings, so we would need
to remove the option for tablets in settings, because currently, the option is there and doesn't actually do anything.

Patch-set: 1
Label: Code-Review=0